windows_agent/requirements.txt
psutil==6.0.0
requests==2.32.3
websockets==12.0
python-dotenv==1.0.1
keyring==25.3.0
PySide6==6.7.2
pywin32==306
